Use .internal_test_app instead of spec/internal for storing the test application
