+++ lib/engine_cart.rb
@@ -4,38 +4,7 @@ require 'bundler'
 
 module EngineCart
   require "engine_cart/engine" if defined? Rails
-
-  class << self
-
-    ##
-    # Name of the engine we're testing
-    attr_accessor :engine_name
-
-    ##
-    # Destination to generate the test app into
-    attr_accessor :destination
-
-    ##
-    # Path to a Rails application template
-    attr_accessor :template
-
-    ##
-    # Path to test app templates to make available to
-    # the test app generator
-    attr_accessor :templates_path
-
-
-    ##
-    # Additional options when generating a test rails application
-    attr_accessor :rails_options
-
-  end
-
-  self.engine_name = ENV["CURRENT_ENGINE_NAME"]
-  self.destination = ENV['ENGINE_CART_DESTINATION'] || ENV['RAILS_ROOT'] || "./spec/internal"
-  self.template = ENV["ENGINE_CART_TEMPLATE"]
-  self.templates_path = ENV['ENGINE_CART_TEMPLATES_PATH'] || "./spec/test_app_templates"
-  self.rails_options = ENV['ENGINE_CART_RAILS_OPTIONS']
+  require 'engine_cart/params'
 
   def self.current_engine_name
     engine_name || File.basename(Dir.glob("*.gemspec").first, '.gemspec')
