+++ lib/engine_cart/tasks/engine_cart.rake
@@ -47,13 +47,14 @@ namespace :engine_cart do
         has_web_console = false
 
         # Hack for https://github.com/rails/web-console/issues/150
-        IO.write("spec/internal/Gemfile", File.open("spec/internal/Gemfile") do |f|
+        gemfile = File.join(EngineCart.destination, 'Gemfile')
+        IO.write(gemfile, File.open(gemfile) do |f|
           text = f.read
           has_web_console = text.match('web-console')
           text.gsub(/.*web-console.*/, "").gsub(/.*IRB console on exception pages.*/, "")
         end)
 
-        File.open("spec/internal/Gemfile", "a") do |f|
+        File.open(gemfile, "a") do |f|
           f.puts 'gem "web-console", group: :development'
         end if has_web_console
       end
