+++ spec/integration/engine_cart_spec.rb
@@ -1,14 +1,15 @@
 require 'spec_helper'
 
 describe "EngineCart powered application" do
-  TEST_APP = File.expand_path("../internal", File.dirname(__FILE__))
+  EngineCart.destination = File.expand_path("../../.internal_test_gem", File.dirname(__FILE__))
+
   it "should have the test_app_templates pre-generated" do
-    expect(File).to exist File.expand_path("spec/test_app_templates", TEST_APP)
+    expect(File).to exist File.expand_path("spec/test_app_templates", EngineCart.destination)
   end
 
   it "should ignore the test app" do
-    git_ignore = File.expand_path(".gitignore", TEST_APP)
-    expect(File.read(git_ignore)).to match /spec\/internal/
+    git_ignore = File.expand_path(".gitignore", EngineCart.destination)
+    expect(File.read(git_ignore)).to match /.internal_test_app/
   end
 
   it "should have a engine_cart:generate rake task available" do
@@ -28,27 +29,27 @@ describe "EngineCart powered application" do
   it "should create a rails app when the engine_cart:generate task is invoked" do
     EngineCart.within_test_app do
       `rake engine_cart:generate`
-      expect(File).to exist(File.expand_path(EngineCart.destination))
+      expect(File).to exist(File.expand_path('.internal_test_app'))
     end
   end
 
   it "should not recreate an existing rails app when the engine_cart:generate task is reinvoked" do
     EngineCart.within_test_app do
       `rake engine_cart:generate`
-      expect(File).to exist(File.expand_path(EngineCart.destination))
-      FileUtils.touch File.join(EngineCart.destination, ".this_should_still_exist")
+      expect(File).to exist(File.expand_path('.internal_test_app'))
+      FileUtils.touch File.join('.internal_test_app', ".this_should_still_exist")
       `rake engine_cart:generate`
-      expect(File).to exist(File.expand_path(File.join(EngineCart.destination, ".this_should_still_exist")))
+      expect(File).to exist(File.expand_path(File.join('.internal_test_app', ".this_should_still_exist")))
     end
   end
   
   it "should create a rails app when the fingerprint argument is changed" do
     EngineCart.within_test_app do
       `rake engine_cart:generate[this-fingerprint]`
-      expect(File).to exist(File.expand_path(EngineCart.destination))
-      FileUtils.touch File.join(EngineCart.destination, ".this_should_not_exist")
+      expect(File).to exist(File.expand_path('.internal_test_app'))
+      FileUtils.touch File.join('.internal_test_app', ".this_should_not_exist")
       `rake engine_cart:generate[that-fingerprint]`
-      expect(File).to_not exist(File.expand_path(File.join(EngineCart.destination, ".this_should_not_exist")))
+      expect(File).to_not exist(File.expand_path(File.join('.internal_test_app', ".this_should_not_exist")))
     end
   end
